% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ednacompare.R
\name{compare_datasets}
\alias{compare_datasets}
\title{Compare datasets}
\usage{
compare_datasets(ds_1, ds_2, labels = c("1", "2"))
}
\arguments{
\item{ds_1}{First dataset}

\item{ds_2}{Second dataset}

\item{labels}{Dataset labels}
}
\value{
A list
}
\description{
Compare datasets
}
