% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ednacompare.R
\name{create_amplicon_length_plot}
\alias{create_amplicon_length_plot}
\title{Create amplicon length plot}
\usage{
create_amplicon_length_plot(ds_1, ds_2, labels)
}
\arguments{
\item{ds_1}{First dataset}

\item{ds_2}{Second dataset}

\item{labels}{Dataset labels}
}
\value{
A ggplot object
}
\description{
Create amplicon length plot
}
