% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_code.R
\name{create_asv_table_reactable}
\alias{create_asv_table_reactable}
\title{Create ASV table}
\usage{
create_asv_table_reactable(ds_1, ds_2, labels)
}
\arguments{
\item{ds_1}{First dataset}

\item{ds_2}{Second dataset}

\item{labels}{Dataset labels}
}
\value{
A reactable table
}
\description{
Create ASV table
}
