% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_code.R
\docType{package}
\name{ednacompare}
\alias{ednacompare-package}
\alias{ednacompare}
\title{ednacompare}
\description{
Compare DwC formatted eDNA datasets.
}
\author{
\strong{Maintainer}: Pieter Provoost \email{pieterprovoost@gmail.com}

}
