% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_code.R
\name{generate_comparison_report}
\alias{generate_comparison_report}
\title{Generate comparison report}
\usage{
generate_comparison_report(results, output_file = "output.html")
}
\arguments{
\item{results}{Results list}

\item{output_file}{Output filename}
}
\value{
None
}
\description{
Generate comparison report
}
