% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_code.R
\name{read_dataset}
\alias{read_dataset}
\title{Read dataset}
\usage{
read_dataset(dataset_path)
}
\arguments{
\item{dataset_path}{Path to the dataset directory, should contain occurrence.txt and dnaderiveddata.txt}
}
\value{
Data frame
}
\description{
Read dataset
}
