% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ednacompare.R
\name{read_dataset}
\alias{read_dataset}
\title{Read dataset}
\usage{
read_dataset(
  dataset_path,
  occurrence_file = "occurrence.txt",
  dna_file = "dnaderiveddata.txt"
)
}
\arguments{
\item{dataset_path}{Path to the dataset directory, should contain occurrence.txt and dnaderiveddata.txt}

\item{occurrence_file}{Occurrence filename}

\item{dna_file}{DNADerivedData filename}
}
\value{
Data frame
}
\description{
Read dataset
}
